structured classes and packages
