+++ executor-service/src/main/java/executor/service/execution/ExecutionServiceImpl.java
@@ -1,6 +1,6 @@
-package executor.service.exectuion;
+package executor.service.execution;
 
-import executor.service.exectuion.scenario.ScenarioExecutor;
+import executor.service.execution.scenario.ScenarioExecutor;
 import executor.service.maintenance.ScenarioSourceListener;
 import executor.service.model.ScenarioDto;
 import org.openqa.selenium.WebDriver;
