+++ executor-service/src/main/java/executor/service/execution/ParallelFlowExecutorImpl.java
@@ -1,8 +1,8 @@
-package executor.service.exectuion;
+package executor.service.execution;
 
 import executor.service.factory.webdriverinitializer.WebDriverProvider;
 
-import executor.service.exectuion.scenario.ScenarioExecutor;
+import executor.service.execution.scenario.ScenarioExecutor;
 import executor.service.maintenance.ScenarioSourceListener;
 import executor.service.maintenance.plugin.proxy.ProxySourcesClient;
