+++ executor-service/src/main/java/executor/service/execution/scenario/ScenarioExecutor.java
@@ -1,4 +1,4 @@
-package executor.service.exectuion.scenario;
+package executor.service.execution.scenario;
 
 import executor.service.model.ScenarioDto;
 import org.openqa.selenium.WebDriver;
