+++ executor-service/src/main/java/executor/service/execution/scenario/step/ClickXpath.java
@@ -1,26 +1,26 @@
-package executor.service.exectuion.scenario.step;
+package executor.service.execution.scenario.step;
 
 import executor.service.annotation.Logged;
-import executor.service.exception.scenario.step.ClickCssException;
+import executor.service.exception.scenario.step.ClickXPathException;
 import executor.service.model.StepDto;
 import org.openqa.selenium.*;
 import org.springframework.stereotype.Component;
 
 @Component
 @Logged
-public class ClickCss implements StepExecution {
+public class ClickXpath implements StepExecution {
     @Override
     public String getStepAction() {
-        return "clickCss";
+        return "clickXpath";
     }
 
     @Override
     public void step(WebDriver webDriver, StepDto stepDto) {
         try {
-            webDriver.findElement(By.cssSelector(stepDto.getValue())).click();
+            webDriver.findElement(By.xpath(stepDto.getValue())).click();
         } catch (NoSuchElementException | ElementNotInteractableException
                  | StaleElementReferenceException | TimeoutException | InvalidSelectorException ex) {
-            throw new ClickCssException(ex);
+            throw new ClickXPathException(ex);
         }
     }
 }
