+++ executor-service/src/main/java/executor/service/logger/ScenarioExceptionHandlingAspect.java
@@ -16,7 +16,7 @@ public class ScenarioExceptionHandlingAspect {
         this.logger = logger;
     }
 
-    @Around("execution(* executor.service.exectuion.scenario.ScenarioExecutor+.*(..))")
+    @Around("execution(* executor.service.execution.scenario.ScenarioExecutor+.*(..))")
     public void handle(ProceedingJoinPoint joinPoint) throws Throwable {
         try {
             joinPoint.proceed();
