+++ executor-service/src/test/java/executor/service/execution/ParallelFlowExecutorImplTest.java
@@ -1,9 +1,7 @@
 package executor.service.execution;
 
 import executor.service.factory.webdriverinitializer.WebDriverProvider;
-import executor.service.exectuion.ExecutionService;
-import executor.service.exectuion.ParallelFlowExecutorImpl;
-import executor.service.exectuion.scenario.ScenarioExecutor;
+import executor.service.execution.scenario.ScenarioExecutor;
 import executor.service.maintenance.ScenarioSourceListener;
 import executor.service.maintenance.plugin.proxy.ProxySourcesClient;
 import executor.service.model.ProxyConfigHolderDto;
