+++ executor-service/src/test/java/executor/service/execution/scenario/ScenarioExecutorImplTest.java
@@ -1,11 +1,10 @@
 package executor.service.execution.scenario;
 
-import executor.service.exectuion.scenario.ScenarioExecutorImpl;
 import executor.service.model.ScenarioDto;
-import executor.service.exectuion.scenario.step.ClickCss;
-import executor.service.exectuion.scenario.step.ClickXpath;
-import executor.service.exectuion.scenario.step.Sleep;
-import executor.service.exectuion.scenario.step.StepExecution;
+import executor.service.execution.scenario.step.ClickCss;
+import executor.service.execution.scenario.step.ClickXpath;
+import executor.service.execution.scenario.step.Sleep;
+import executor.service.execution.scenario.step.StepExecution;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
