+++ executor-service/src/test/java/executor/service/execution/scenario/step/ClickXpathTest.java
@@ -1,7 +1,6 @@
 package executor.service.execution.scenario.step;
 
 import executor.service.exception.scenario.step.ClickXPathException;
-import executor.service.exectuion.scenario.step.ClickXpath;
 import executor.service.model.StepDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
